+++ R/ever_filter.R
@@ -1,4 +1,4 @@
-#' @importFrom dplyr %>%  filter ungroup is.tbl
+#' @importFrom dplyr filter ungroup is.tbl
 #' @importFrom stringr str_c str_detect
 #' @importFrom tidyr drop_na
 #' @importFrom wrapr stop_if_dot_args let
@@ -10,48 +10,52 @@ NULL
 #' programs.
 #'
 #' To use this function, the \code{midfielddata} package must be installed to
-#' provide \code{midfieldterms}, the default reference data set.
+#' provide the \code{midfieldterms} data frame as the default value for the
+#' \code{data} argument.
 #'
-#' The \code{series} argument is an atomic character vector of 6-digit CIP
-#' codes used to filter the reference data set and extract student IDs
-#' and terms.
+#' A substitute for \code{midfieldterms} may be used if it has the same structure
+#' as \code{midfieldterms}.
 #'
-#' The function returns a subset of \code{reference} with every unique
+#' The \code{filter_by} argument is an atomic character vector of 6-digit CIP
+#' codes used to filter \code{data} and extract student IDs and terms.
+#'
+#' The function returns a subset of \code{data} with every unique
 #' combination of student ID and program CIP, representing every student
 #' ever enrolled in the specified programs. Only the student ID and program
-#' CIP are returned; other variables in \code{reference} are quietly dropped.
+#' CIP are returned; other variables in \code{data} are quietly dropped.
 #'
-#' Optional arguments. An alternate reference data set can be assigned via
-#' the \code{reference} argument. The alternate must have variables for
-#' student ID, 6-digit CIP, and terms with the same structure as in
-#' \code{midfieldterms}. Use the  \code{id}, \code{cip6}, and \code{term}
+#' Optional arguments. Use the  \code{id}, \code{cip6}, and \code{term}
 #' arguments to reassign the variables names, if necessary, to match the
 #' variable names in the alternate.
 #'
-#' @param series atomic character vector of 6-digit CIP codes specifying the
-#' programs to filter by
+#' @param data Data frame of student IDs, academic terms, and CIP codes,
+#' default \code{midfieldterms}.
 #'
-#' @param ... not used for values, forces later arguments to bind by name
+#' @param filter_by Atomic character vector of 6-digit CIP codes specifying the
+#' programs to filter by.
 #'
-#' @param reference a reference data frame from which student IDs, academic
-#' terms, and CIP codes are obtained, default \code{midfieldterms}
+#' @param ... Not used for values, forces later arguments to bind by name
 #'
-#' @param id character column name of the ID variable in \code{reference}
+#' @param id Optional argument, the quoted column name of the student ID
+#' variable in \code{data}. Default is "id".
 #'
-#' @param cip6 character column name of the CIP code variable
-#' in \code{reference}
+#' @param cip6 Optional argument, the quoted column name of the 6-digit CIP
+#' code variable in \code{data}. Default is "cip6".
 #'
-#' @param term character column name of term variable in \code{reference}
+#' @param term Optional argument, the quoted column name of the term
+#' variable in \code{data}. Default is "term".
 #'
 #' @return Data frame with character variables for student ID and program CIP
 #'
-#' @seealso \code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
+#' @seealso \code{\link[midfieldr]{cip6_select}} for selecting 6-digit CIP codes
+#' and naming programs.
 #'
 #' @examples
-#' ever_filter(series = "540104")
+#' library("midfielddata")
+#' (ever_filter(filter_by = "540104"))
+#'
 #' @export
-ever_filter <- function(series, ...,
-                        reference = NULL,
+ever_filter <- function(data = NULL, filter_by = NULL, ...,
                         id = "id",
                         cip6 = "cip6",
                         term = "term") {
@@ -65,28 +69,24 @@ ever_filter <- function(series, ...,
   # force optional arguments to be usable only by name
   wrapr::stop_if_dot_args(substitute(list(...)), "ever_filter")
 
-  if (is.null(series)) {
-    stop("midfieldr::ever_filter, series cannot be NULL")
+  if (is.null(data)) {
+    data <- midfielddata::midfieldterms
   }
-
-  if (isFALSE(is.atomic(series))) {
-    stop("midfieldr::ever_filter, series must be an atomic variable")
+  if (!(is.data.frame(data) || dplyr::is.tbl(data))) {
+    stop("midfieldr::ever_filter, data must be a data frame or tbl")
   }
-
-  # assign the default terms dataset
-  if (is.null(reference)) {
-    reference <- midfielddata::midfieldterms
+  if (is.null(filter_by)) {
+    stop("midfieldr::ever_filter, filter_by cannot be NULL")
   }
-
-  if (!(is.data.frame(reference) || dplyr::is.tbl(reference))) {
-    stop("midfieldr::ever_filter, reference must be a data frame or tbl")
+  if (isFALSE(is.atomic(filter_by))) {
+    stop("midfieldr::ever_filter, filter_by must be an atomic variable")
   }
 
   # search terms in a single string
-  collapse_series <- stringr::str_c(series, collapse = "|")
+  collapse_series <- stringr::str_c(filter_by, collapse = "|")
 
   # addresses R CMD check warning "no visible binding"
-  ID <- NULL
+  ID   <- NULL
   CIP6 <- NULL
   TERM <- NULL
 
@@ -95,8 +95,8 @@ ever_filter <- function(series, ...,
   wrapr::let(
     alias = mapping,
     expr = {
-      # filter for data for specified programs
-      students <- dplyr::select(reference, ID, CIP6, TERM)
+      # filter for specified programs
+      students <- dplyr::select(data, ID, CIP6, TERM)
       students <- dplyr::filter(
         students,
         stringr::str_detect(CIP6, collapse_series)
@@ -108,7 +108,7 @@ ever_filter <- function(series, ...,
       students <- dplyr::filter(students, dplyr::row_number() == 1)
       students <- dplyr::ungroup(students)
 
-      # clean up before return (dop_na and unique just in case)
+      # clean up before return (drop_na and unique just in case)
       students <- dplyr::select(students, ID, CIP6)
       students <- tidyr::drop_na(students)
       students <- unique(students)