+++ vignettes/stickiness.R
@@ -1,0 +1,4 @@
+## ----opts.label = 'dopurl'----------------------------------------------------
+students <- ever_filter(data = midfieldterms, filter_by = program_series) %>% 
+        print()
+